% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_channel_noise.R
\name{add_channel_noise}
\alias{add_channel_noise}
\title{Add channel noise}
\usage{
add_channel_noise(codebook, signal, noise_power, timesteps)
}
\arguments{
\item{codebook}{constructed using construct_codebook function}

\item{signal}{data_frame output from transmit_signal, that has spikes allocated by signal_power}

\item{noise_power}{spikes per second to add to every Poisson process, i.e. for each symbol in the codebook}

\item{timesteps}{length of time to simulate. needed to know how much noise to produce.}
}
\value{
data frame with noise appended
}
\description{
Takes a list of 'signal' spikes and adds noise spikes
}
\examples{

}
