% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_codebook.R
\name{construct_codebook}
\alias{construct_codebook}
\title{Construct a codebook object that maps symbols to neurons}
\usage{
construct_codebook(symbols, neuron_allocation = c())
}
\arguments{
\item{symbols}{list of strings, representing symbols to be encoded}

\item{neuron_allocation}{a vector of counts, i.e. how many neurons to associate with each symbol.}
}
\value{

}
\description{
Some other stuff here.
}
\examples{

}
