% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{decode_signal}
\alias{decode_signal}
\title{Determine which symbol is being transmitted.}
\usage{
decode_signal(codebook, signal_plus_noise, signal_power, noise_power,
  prior_distribution = c(), time_interval_in_seconds = 0.1,
  entropy_threshold = 0.1, return_entropy_trace = FALSE,
  return_posterior_trace = FALSE,
  return_posterior_at_stop_time = FALSE)
}
\arguments{
\item{codebook}{constructed using construct_codebook function}

\item{signal_plus_noise}{signal to be decoded, output by add_channel_noise}

\item{signal_power}{spikes per second allocated to signal process}

\item{noise_power}{spikes per second allocated to every process (including signal process)}

\item{prior_distribution}{initial guess at the distribution of messages. Influences decoding times.}

\item{time_interval_in_seconds}{the interval at which to perform inference}

\item{entropy_threshold}{transmission stops after entropy reaches this threshold}

\item{return_entropy_trace}{return posterior entropy for each time step}

\item{return_posterior_trace}{return trace of posterior values for every time step}

\item{return_posterior_at_stop_time}{returns posterior values at stop time}
}
\value{
list of data frames. Will list decoded symbol, stop_time (in units of interval), entropy threshold, posterior at stop time, and potentially other time series as specified in the function call
}
\description{
Decoder takes a list of spike times for each Poisson process.  It knows that one of them is 
firing at a higher rate, and tries to determine which it is by comparing likelihoods of observations
}
\examples{

}
