% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{encode_signal}
\alias{encode_signal}
\title{Encode signal into Poisson spikes}
\usage{
encode_signal(codebook, symbol, signal_power, duration_in_seconds,
  leak_symbols = NA, leak_powers = NA)
}
\arguments{
\item{codebook}{constructed using construct_codebook function}

\item{symbol}{symbol to be encoded}

\item{signal_power}{spikes per second allocated to signal process}

\item{duration_in_seconds}{number of time (in seconds) to simulate for -- limits max decoding time. Signals not decoded by this time will be interpreted as 'NA'}

\item{leak_symbols}{vector of symbols to which a 'leak' signal power should be added. Should not include the symbol specified by the 'symbol' parameter}

\item{leak_powers}{vector of signal powers (spikes/second), to assign to leak symbols. Must be same length as leak_symbols.}
}
\value{

}
\description{
The 'code' is just a Poisson spike train for a given group of neurons
This function will return a data frame with (group_index,spike_time) columns
}
\details{
Some other stuff here.
}
\examples{

}
