% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_entropy_decrease_trace.R
\name{plot_entropy_decrease_trace}
\alias{plot_entropy_decrease_trace}
\title{Plots spikes and an entropy trace for a single transmitted signal.}
\usage{
plot_entropy_decrease_trace(symbols = c("A", "B", "C", "D"),
  symbol_to_transmit = "A", signal_power = 3, noise_power = 3,
  max_stop_time_in_seconds = 5, min_stop_time_in_seconds = 3,
  duration_in_seconds = 50, entropy_threshold = 0.3, file_path = NA)
}
\arguments{
\item{symbols}{vector of character symbols}

\item{symbol_to_transmit}{Symbol that signal power should be allocated to}

\item{signal_power}{spikes per second for signal power}

\item{noise_power}{spikes per second for noise power}

\item{max_stop_time_in_seconds}{search until the stop_time is between max and min}

\item{min_stop_time_in_seconds}{search until the stop_time is between max and min}

\item{duration_in_seconds}{seconds to simulate for - should be highish}

\item{entropy_threshold}{stopping_threshold. Should be > log2(length(symbols))}

\item{file_path}{optional file path for saving the generated lot}
}
\value{
list with three elements, (1) signal_plus_noise, (2) decoded signal with entropy trace, (3) generated plot
}
\description{
Optionally save the result to file.
}
