% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_experiment.R
\name{prior_experiment}
\alias{prior_experiment}
\title{Run a bunch of simulations with a given}
\usage{
prior_experiment(codebook, source_distribution, prior_distribution,
  threshold_type, threshold_value, signal_power, noise_power, timesteps,
  time_interval, repeats)
}
\arguments{
\item{codebook}{constructed using construct_codebook function}

\item{source_distribution}{source distribution of messages. Messages are generated in simulations according to this distribution.}

\item{prior_distribution}{initial guess at the distribution of messages. Influences decoding times.}

\item{signal_power}{spikes per second allocated to signal process}

\item{noise_power}{spikes per second allocated to every process (including signal process)}

\item{timesteps}{number of time (in seconds) to simulate for -- limits max decoding time. Signals not decoded by this time will be interpreted as 'NA'}

\item{time_interval}{time resolution for simulations (like dt)}

\item{repeats}{number of times to run the simulation}

\item{entropy_threshold}{transmission stops after entropy reaches this threshold}
}
\value{
data frame containing encoded and decoded symbols
}
\description{
Some other stuff here.
}
\examples{

}
