% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmit_signal.R
\name{transmit_signal}
\alias{transmit_signal}
\title{Encode the signal, add noise, and decode the signal}
\usage{
transmit_signal(codebook, symbol, signal_power, noise_power, timesteps,
  prior, time_interval = 0.1, threshold = "entropy",
  threshold_value = 0.1, return_posteriors = FALSE,
  leak_symbols = NA, leak_powers = NA)
}
\arguments{
\item{codebook}{constructed using construct_codebook function}

\item{symbol}{symbol to transmit. Must be defined in the codebook.}

\item{signal_power}{spikes per second allocated to signal process}

\item{noise_power}{spikes per second allocated to every process (including signal process)}

\item{timesteps}{number of time (in seconds) to simulate for -- limits max decoding time. Signals not decoded by this time will be interpreted as 'NA'}

\item{time_interval}{time resolution for simulations (like dt)}

\item{prior_distribution}{initial guess at the distribution of messages. Influences decoding times.}

\item{entropy_threshold}{transmission stops after entropy reaches this threshold}

\item{repeats}{number of times to run the simulation}
}
\value{
returns which symbol is decided plus response time and which symbol was sent
}
\description{
Helper function to encapsulate function calls for a 'signal transmission' event
}
\examples{

}
