% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{add_channel_noise}
\alias{add_channel_noise}
\title{Add channel noise}
\usage{
add_channel_noise(codebook, signal, noise_power, duration_in_seconds)
}
\arguments{
\item{codebook}{constructed using construct_codebook function}

\item{signal}{data_frame output from transmit_signal, that has spikes allocated by signal_power}

\item{noise_power}{spikes per second to add to every Poisson process, i.e. for each symbol in the codebook}

\item{duration_in_seconds}{length of time to simulate, in seconds. needed to know how much noise to produce.}
}
\value{
data_frame with (group_index,spike_times) columns, representing the signal with noise appended
}
\description{
Takes a list of 'signal' spikes and adds noise spikes
}
\examples{

}
