% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{construct_codebook}
\alias{construct_codebook}
\title{Construct a codebook object that maps symbols to neurons}
\usage{
construct_codebook(symbols)
}
\arguments{
\item{symbols}{list of strings, representing symbols to be encoded}
}
\value{
list of entries, where each entry details a symbol and the neuron/neural group that is allocated signal power when transmitted
}
\description{
Some other stuff here.
}
\examples{

}
