% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_signal.R
\name{encode_signal}
\alias{encode_signal}
\title{Encode signal into Poisson spikes}
\usage{
encode_signal(codebook, symbol, signal_power, timesteps,
  leak_symbols = NA, leak_powers = NA)
}
\arguments{
\item{codebook}{constructed using construct_codebook function}

\item{symbol}{symbol to be encoded}

\item{signal_power}{spikes per second allocated to signal process}

\item{timesteps}{number of time (in seconds) to simulate for -- limits max decoding time. Signals not decoded by this time will be interpreted as 'NA'}

\item{leak_symbols}{}

\item{leak_powers}{}
}
\value{

}
\description{
The 'code' is just a Poisson spike train for a given group of neurons
This function will return a data frame with (group_index,spike_time) columns
}
\details{
Some other stuff here.
}
\examples{

}
